requirements.txt
streamlit
transformers
torch
opencv-python
Pillow
deepface
sentence-transformers
scikit-learn
tqdm
numpy
pytesseract
